# backend/docs/openapi.v0.1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "duoduo-product-testing-api",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/products/upload-url": {
      "post": {
        "tags": [
          "products"
        ],
        "summary": "Create Product Upload Url",
        "description": "Return a mock upload URL for a product image.",
        "operationId": "create_product_upload_url_api_v1_products_upload_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUploadUrlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductUploadUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/products": {
      "post": {
        "tags": [
          "products"
        ],
        "summary": "Create Product",
        "description": "Create a product with mock AI understanding.",
        "operationId": "create_product_api_v1_products_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "products"
        ],
        "summary": "List Products",
        "description": "List the current user's products.",
        "operationId": "list_products_api_v1_products_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{product_id}": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Get Product",
        "description": "Return one product owned by the current user.",
        "operationId": "get_product_api_v1_products__product_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/products/{product_id}/reanalyze": {
      "post": {
        "tags": [
          "products"
        ],
        "summary": "Reanalyze Product",
        "description": "Refresh mock product understanding.",
        "operationId": "reanalyze_product_api_v1_products__product_id__reanalyze_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/surveys/generate": {
      "post": {
        "tags": [
          "surveys"
        ],
        "summary": "Generate Survey",
        "description": "Generate a mock AI survey from the local seed template.",
        "operationId": "generate_survey_api_v1_surveys_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SurveyGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/surveys/{survey_id}": {
      "get": {
        "tags": [
          "surveys"
        ],
        "summary": "Get Survey",
        "description": "Return one survey owned through the current user's evaluation.",
        "operationId": "get_survey_api_v1_surveys__survey_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "survey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Survey Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/surveys/{survey_id}/questions": {
      "put": {
        "tags": [
          "surveys"
        ],
        "summary": "Update Survey Questions",
        "description": "Overwrite survey questions while the evaluation is editable.",
        "operationId": "update_survey_questions_api_v1_surveys__survey_id__questions_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "survey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Survey Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/SurveyQuestion"
                },
                "title": "Questions"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProductAiSummary": {
        "properties": {
          "main_selling_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Main Selling Points"
          },
          "key_ingredients": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Key Ingredients"
          },
          "suitable_skin_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Suitable Skin Types"
          },
          "target_audience": {
            "type": "string",
            "title": "Target Audience",
            "default": ""
          },
          "competitive_position": {
            "type": "string",
            "title": "Competitive Position",
            "default": ""
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "sub_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Category"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "price_range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Range"
          },
          "target_channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Channel"
          },
          "claims_detected": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Claims Detected"
          },
          "risk_or_uncertainty_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Risk Or Uncertainty Points"
          },
          "usage_scenarios": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Usage Scenarios"
          },
          "questionnaire_focus": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Questionnaire Focus"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          }
        },
        "type": "object",
        "title": "ProductAiSummary",
        "description": "Structured product understanding from AI or mock."
      },
      "ProductCreateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "minLength": 10,
            "title": "Description"
          },
          "image_object_keys": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Image Object Keys"
          },
          "image_base64_list": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 5
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Base64 List",
            "description": "Base64-encoded images (raw base64, data-URL, or HTTPS URL). Max 5 images. Each string must be under 8 MB."
          },
          "brand": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "target_channel": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Channel"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "description"
        ],
        "title": "ProductCreateRequest",
        "description": "Product creation request.\n\nImages can be supplied in two ways (not mutually exclusive):\n- ``image_object_keys``: object keys from a prior upload-url flow\n- ``image_base64_list``: raw base64 strings or data-URLs sent directly\n\nAt least one of the two must be non-empty."
      },
      "ProductListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProductResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "next_cursor",
          "has_more"
        ],
        "title": "ProductListResponse",
        "description": "Cursor-paginated product list response."
      },
      "ProductResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "image_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Image Urls"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "sub_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sub Category"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "price_range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Range"
          },
          "target_channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Channel"
          },
          "ai_summary": {
            "$ref": "#/components/schemas/ProductAiSummary"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "image_urls",
          "category",
          "sub_category",
          "brand",
          "price",
          "price_range",
          "target_channel",
          "ai_summary",
          "status",
          "created_at"
        ],
        "title": "ProductResponse",
        "description": "Product response aligned to the API contract."
      },
      "ProductUploadUrlRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "size_bytes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Size Bytes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "filename",
          "mime_type",
          "size_bytes"
        ],
        "title": "ProductUploadUrlRequest",
        "description": "Product image upload URL request."
      },
      "ProductUploadUrlResponse": {
        "properties": {
          "upload_url": {
            "type": "string",
            "title": "Upload Url"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Headers"
          },
          "object_key": {
            "type": "string",
            "title": "Object Key"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "upload_url",
          "method",
          "headers",
          "object_key",
          "expires_in"
        ],
        "title": "ProductUploadUrlResponse",
        "description": "Mock product image upload URL response."
      },
...
      "SurveyGenerateRequest": {
        "properties": {
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "evaluation_id": {
            "type": "string",
            "title": "Evaluation Id"
          },
          "extra_focus": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Focus"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "product_id",
          "evaluation_id"
        ],
        "title": "SurveyGenerateRequest",
        "description": "Generate survey request."
      },
      "SurveyQuestion": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Id"
          },
          "dim": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Dim"
          },
          "type": {
            "type": "string",
            "enum": [
              "single",
              "multi",
              "scale_1_5",
              "open"
            ],
            "title": "Type"
          },
          "question": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Question"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "id",
          "dim",
          "type",
          "question"
        ],
        "title": "SurveyQuestion",
        "description": "Survey question shape from the API contract."
      },
      "SurveyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "evaluation_id": {
            "type": "string",
            "title": "Evaluation Id"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "generated_by": {
            "type": "string",
            "title": "Generated By"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/SurveyQuestion"
            },
            "type": "array",
            "title": "Questions"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "evaluation_id",
          "product_id",
          "version",
          "generated_by",
          "questions",
          "created_at"
        ],
        "title": "SurveyResponse",
        "description": "Survey response aligned to the API contract."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}